feat: Repo 확장성 고려한 설계 변경
